Fix squid:S1214: Constants should not be defined in interfaces Move UserAccountService.MIN_PASSWORD_LENGTH to UserAccountConstants.MIN_PASSWORD_LENGTH
