+++ molgenis-security/src/main/java/org/molgenis/security/user/UserAccountConstants.java
@@ -1,0 +1,8 @@
+package org.molgenis.security.user;
+
+public final class UserAccountConstants {
+
+  public static final int MIN_PASSWORD_LENGTH = 6;
+
+  private UserAccountConstants() {}
+}
